+++ types/base.d.ts
@@ -2,6 +2,7 @@ interface ParameterConfiguration {
   index: number;
   type: any;
   name?: string;
+  data?: any;
 }
 
 interface Params {